@@ -16,13 +16,13 @@
 
 package com.example.android.teatime;
 
-import static org.junit.Assert.assertEquals;
+// import static org.junit.Assert.assertEquals;
 
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+// import android.content.Context;
+// import androidx.test.InstrumentationRegistry;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import org.junit.Test;
+// import org.junit.Test;
 import org.junit.runner.RunWith;
 
 /**
@@ -30,6 +30,7 @@ import org.junit.runner.RunWith;
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
+@RunWith(AndroidJUnit4.class)
 public class ExampleInstrumentedTest {
 
 }
