@@ -18,29 +18,32 @@ package com.example.android.teatime;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.GridView;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
+import com.example.android.teatime.databinding.ActivityMenuBinding;
 import com.example.android.teatime.model.Tea;
 
 import java.util.ArrayList;
 
 public class MenuActivity extends AppCompatActivity {
 
-    Intent mTeaIntent;
+    private Intent mTeaIntent;
+    private ActivityMenuBinding binding;
 
     public final static String EXTRA_TEA_NAME = "com.example.android.teatime.EXTRA_TEA_NAME";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_menu);
-        Toolbar menuToolbar = (Toolbar) findViewById(R.id.menu_toolbar);
-        setSupportActionBar(menuToolbar);
-        getSupportActionBar().setTitle(getString(R.string.menu_title));
+        binding = ActivityMenuBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
+        setSupportActionBar(binding.menuToolbar);
+        if (getSupportActionBar() != null) getSupportActionBar().setTitle(getString(R.string.menu_title));
 
         // Create an ArrayList of teas
         final ArrayList<Tea> teas = new ArrayList<>();
@@ -53,26 +56,19 @@ public class MenuActivity extends AppCompatActivity {
 
         // Create a {@link TeaAdapter}, whose data source is a list of {@link Tea}s.
         // The adapter know how to create grid items for each item in the list.
-        GridView gridview = (GridView) findViewById(R.id.tea_grid_view);
         TeaMenuAdapter adapter = new TeaMenuAdapter(this, R.layout.grid_item_layout, teas);
-        gridview.setAdapter(adapter);
+        binding.teaGridView.setAdapter(adapter);
 
 
         // Set a click listener on that View
-        gridview.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
+        binding.teaGridView.setOnItemClickListener((adapterView, view, position, id) -> {
 
-                Tea item = (Tea) adapterView.getItemAtPosition(position);
-                // Set the intent to open the {@link OrderActivity}
-                mTeaIntent = new Intent(MenuActivity.this, OrderActivity.class);
-                String teaName = item.getTeaName();
-                mTeaIntent.putExtra(EXTRA_TEA_NAME, teaName);
-                startActivity(mTeaIntent);
-
-            }
+            Tea item = (Tea) adapterView.getItemAtPosition(position);
+            // Set the intent to open the {@link OrderActivity}
+            mTeaIntent = new Intent(MenuActivity.this, OrderActivity.class);
+            String teaName = item.getTeaName();
+            mTeaIntent.putExtra(EXTRA_TEA_NAME, teaName);
+            startActivity(mTeaIntent);
         });
-
     }
-
 }
