@@ -25,6 +25,8 @@ import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.annotation.LayoutRes;
+
 import com.example.android.teatime.model.Tea;
 
 import java.util.ArrayList;
@@ -36,11 +38,12 @@ import java.util.ArrayList;
 
 public class TeaMenuAdapter extends ArrayAdapter<Tea> {
 
-    private Context mContext;
-    private int layoutResourceId;
-    private ArrayList data = new ArrayList();
+    @LayoutRes
+    private final int layoutResourceId;
+    private final Context mContext;
+    private ArrayList<Tea> data;
 
-    public TeaMenuAdapter(Context context, int layoutResourceId, ArrayList data) {
+    public TeaMenuAdapter(Context context, @LayoutRes int layoutResourceId, ArrayList data) {
         super(context, layoutResourceId, data);
         this.layoutResourceId = layoutResourceId;
         this.mContext = context;
@@ -64,8 +67,8 @@ public class TeaMenuAdapter extends ArrayAdapter<Tea> {
             LayoutInflater inflater = ((Activity) mContext).getLayoutInflater();
             convertView = inflater.inflate(layoutResourceId, parent, false);
             holder = new ViewHolder();
-            holder.imageTitle = (TextView) convertView.findViewById(R.id.tea_grid_name);
-            holder.image = (ImageView) convertView.findViewById(R.id.image);
+            holder.imageTitle = convertView.findViewById(R.id.tea_grid_name);
+            holder.image = convertView.findViewById(R.id.image);
             convertView.setTag(holder);
         } else {
             holder = (ViewHolder) convertView.getTag();;
@@ -75,5 +78,4 @@ public class TeaMenuAdapter extends ArrayAdapter<Tea> {
         holder.image.setImageResource(currentTea.getImageResourceId());
         return convertView;
     }
-
 }
